Massive updates to incorporate new version of colony.
Made great progress.  Not completely done, but I have
dealt with all the figures for the paper.  Still need to do X-plots and running time supplement plots, and I need to make the running times table.  Then have to to the high confidence errors in new version of colony fo the table and we will be ready to do final edits to paper...